feat: execute (#14)
Successfully execute the target example block, including the full flow. When a new payload message arrives, it is first validated, and then, within the fetched witness, it is executed in the EVM. If the bytecode does not exist on disk, fetch it from the network on demand and save it to disk.